requirements.txt
djangorestframework==3.14.0
markdown==3.4.4
Django==4.2.3
django-filter==23.2
